Add sequence detection in parsing trie generation; move parsing trie generation into separate file; add comments explaining Suffix and SuffixClump structs
